+++ sain/cell/once.py
@@ -32,12 +32,13 @@
 
 from __future__ import annotations
 
-__all__ = ("Once",)
+__all__ = ("Once", "AsyncOnce")
 
-import _thread
+import asyncio
+import threading
 import typing
 
-from . import option as _option
+from sain import option as _option
 
 if typing.TYPE_CHECKING:
     from sain import Option
@@ -64,7 +65,7 @@ class Once(typing.Generic[T]):
 
     def spawn(app: Application) -> None:
         app.uuid.get().is_none() # True
-        uid = app.uuid.get_or_init(uuid4())
+        uid = app.uuid.get_or(uuid4())
         app.send(uid)
 
     def run_application():
@@ -77,11 +78,10 @@ class Once(typing.Generic[T]):
     ```
     """
 
-    __slots__ = ("_inner", "_lock", "_blocking")
+    __slots__ = ("_inner", "_lock")
 
-    def __init__(self, *, blocking: bool = True) -> None:
-        self._blocking = blocking
-        self._lock: _thread.LockType | None = None
+    def __init__(self) -> None:
+        self._lock: threading.Lock | None = None
         self._inner: T | None = None
 
     @property
@@ -108,30 +108,28 @@ class Once(typing.Generic[T]):
         if self._inner is not None:
             raise ValueError("Value is already set.")
 
-        self._inner = origin = self.get_or_init(v)
+        self._inner = origin = self.get_or(v)
         self._lock = None
         return origin
 
     def clear(self) -> None:
+        """Clear the inner value, Setting it to `None`."""
         self._lock = None
         self._inner = None
 
-    def get_or_init(self, f: T) -> T:
-        """Get the value if its not `None`, Otherwise call `f()` setting the value and returning it."""
+    def get_or(self, f: T) -> T:
+        """Get the value if its not `None`, Otherwise set `f` value and returning it."""
 
         # If the value is not empty we return it immediately.
         if self._inner is not None:
             return self._inner
 
         if self._lock is None:
-            self._lock = _thread.allocate_lock()
+            self._lock = threading.Lock()
 
-        try:
-            self._lock.acquire(blocking=self._blocking)
+        with self._lock:
             self._inner = f
             return self._inner
-        finally:
-            self._lock.release()
 
     def __repr__(self) -> str:
         return f"Once(value: {self._inner})"
@@ -142,6 +140,112 @@ class Once(typing.Generic[T]):
     def __eq__(self, __value: object) -> bool:
         if isinstance(__value, Once):
             return self._inner == __value._inner  # pyright: ignore[reportUnknownVariableType, reportUnknownMemberType]
+
+        if isinstance(self._inner, type(__value)):
+            return __value == self._inner
+
+        return NotImplemented
+
+    def __ne__(self, __value: object) -> bool:
+        return not self.__eq__(__value)
+
+    def __bool__(self) -> bool:
+        return self.is_set
+
+
+@typing.final
+class AsyncOnce(typing.Generic[T]):
+    """A synchronization primitive which can be written to only once.
+
+    This is an `async` version of `Once`.
+
+    Example
+    -------
+    ```py
+    from sain.once import Once
+    from uuid import uuid4, UUID
+
+    class Application:
+        # Not initialized yet
+        uuid: Once[UUID] = Once()
+
+        async def send(self, id: UUID) -> None:
+            ...
+
+    async def spawn(app: Application) -> None:
+        app.uuid.get().is_none() # True
+        uid = await app.uuid.get_or(uuid4())
+        await app.send(uid)
+
+    def run_application():
+        # This will init the uuid if its not set or return it if it already is.
+        app = Application()
+        tasks = (asyncio.create_task(spawn(app)) for _ in range(2))
+        await asyncio.gather(*tasks)
+
+        app.uuid.get().is_some()  # True
+    ```
+    """
+
+    __slots__ = ("_inner", "_lock")
+
+    def __init__(self) -> None:
+        self._lock: asyncio.Lock | None = None
+        self._inner: T | None = None
+
+    @property
+    def is_set(self) -> bool:
+        return self._inner is not None
+
+    def get(self) -> Option[T]:
+        """Gets the stored value.
+
+        `Option(None)` is returned if nothing is stored. This method will never block.
+        """
+        return _option.Some(self._inner)
+
+    async def set(self, v: T) -> T:
+        """Set the const value if its not set.
+
+        Raises
+        ------
+        `RuntimeError`
+            If the value is already set. This will get raised.
+        """
+        if self._inner is not None:
+            raise ValueError("Value is already set.")
+
+        self._inner = origin = await self.get_or(v)
+        self._lock = None
+        return origin
+
+    def clear(self) -> None:
+        self._lock = None
+        self._inner = None
+
+    async def get_or(self, f: T) -> T:
+        """Get the value if its not `None`, Otherwise set `f` value and returning it."""
+
+        # If the value is not empty we return it immediately.
+        if self._inner is not None:
+            return self._inner
+
+        if self._lock is None:
+            self._lock = asyncio.Lock()
+
+        async with self._lock:
+            self._inner = f
+            return self._inner
+
+    def __repr__(self) -> str:
+        return f"Once(value: {self._inner})"
+
+    def __str__(self) -> str:
+        return str(self._inner)
+
+    def __eq__(self, __value: object) -> bool:
+        if isinstance(__value, Once):
+            return self._inner == __value._inner  # pyright: ignore
 
         if isinstance(self._inner, type(__value)):
             return __value == self._inner